+++ lib/xsd/objects/simple_type.rb
@@ -31,5 +31,23 @@ module XSD
     def linked?
       !name.nil?
     end
+
+    # Get base data type
+    # @return String, nil
+    def data_type
+      if restriction
+        if restriction.base
+          restriction.base_simple_type&.data_type || strip_prefix(restriction.base)
+        else
+          restriction.simple_type&.data_type
+        end
+      elsif union
+        types = union.types.map { |type| type.is_a?(String) ? strip_prefix(type) : type.data_type }.uniq
+        types.size == 1 ? types.first : types
+      else
+        # list is always a sting
+        'string'
+      end
+    end
   end
 end